requirements.txt
azure-functions
flask
flask-mysqldb
pyyaml
